Drop crypto app reliance
By doing this, we fix two issues at once. First, there's no need to have
crypto available anymore. While not having crypto in your Erlang
installation is a questionable packaging decision, it tends to happen in
the wild with users installing Erlang and missing crypto. Sometimes this
is not due to a distro's package but users building Erlang without the
needed OpenSSL dependencies. Second, this resolves the Erlang 20 rng
deprecation warnings.