Improve pseudo-unique ct_run node name
Replacing crypto:rand_uniform(0, 10000) with rebar_rnd:uniform(10000) (which is
either rand:uniform or random:uniform dependent on what otp version it runs on)
fails in rebar_ct. One might assume it's because the old call requests a number
>= 0 while the new call has no way to ask for anything but >= 1. However, the
way the value is used doesn't rely on it to include 0, so the bug lies
elsewhere. Actually, generating a random number to choose a pseudo-unique
ct_run node name isn't a reliable method to avoid naming clashes.

I cannot reproduce it locally and also cannot see the actual error in the CI
logs. Mirroring what rebar_eunit does and using a bigger rng range seems to be
enough to avoid node name clashes on Travis-CI.